google_battery: Fix SSOC delta's misuse of qnum API
qnum_rconst is only meant to be used with compile-time constants because
it relies on compile-time floating-point arithmetic that cannot be done
at runtime in the kernel. qnum_fromint is to be used for non-constant
values instead, but it gives a slightly different result due to not
adding 0.5.

Remove the ability to change ssoc_delta in DT since it's not modified
anyway.

This fixes the following error when compiling with GCC:

../drivers/power/supply/google/google_battery.c: In function 'gbatt_set_property':
../drivers/power/supply/google/google_battery.c:4622:12: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
 4622 | static int gbatt_set_property(struct power_supply *psy,
      |            ^~~~~~~~~~~~~~~~~~

Test: verify value of qnum_rconst(SSOC_DELTA)
Signed-off-by: Danny Lin <danny@kdrag0n.dev>